Fixed smoothing issue in Transfer Function estimation
Smoothing the CSD is required to compute the coherence, but was leading to poor fits for the transfer function. I looked smoothing in polar coordinates, which resulted in less amplitude loss, but messed up the phase. Solution for now is to use un-smoothed CSD to compute the complex transfer response.