Created SSV DMD algorithm, revised SSV additive optimization.
Added a couple SSV tests based on S&P Example 8.9. The Python and MATLAB code produce the same result. The "improved lower bound" using the MU algorithm doesn't work as expected. The DMD "improved upper bound" algorithm matches the result in MATLAB and the text.

The Test for MIMO Ctrl uses a simple QP to solve the additive  uncertainty problem.